% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinypedigree.R
\name{tiny_pedigree}
\alias{tiny_pedigree}
\title{Plot a Pedigree Diagram}
\usage{
tiny_pedigree(
  data,
  links = NULL,
  fold = TRUE,
  symbolsize = 1.8,
  textsize = 0.8
)
}
\arguments{
\item{data}{The main data file with ID,dad,mom,gender and other cols.}

\item{links}{The file with id1 and id2 linking married without kids.}

\item{fold}{Break names at blanks.}

\item{symbolsize}{Adjust the size of all symbols.}

\item{textsize}{Adjust the size of all text.}
}
\value{
A pedigree plot.
}
\description{
Plot a Pedigree Diagram
}
\examples{
tiny_pedigree(data=data,links=links,symbolsize=0.8,textsize=0.7,fold=FALSE)
}
